feat: support SHA2-{224,384} and handle args correctly
Other changes include:
- Handle args correctly:
  - Use correct CLI flag for BSD-style checksums
  - Use clap builtin functions to handle required flags
- Use usize in place of i32 (to avoid unneeded casts)
- Add unit tests for SHA-224 and SHA-384

+++ src/errors.rs
@@ -6,21 +6,21 @@ pub enum ShaSumError {
     #[error(
         "Invalid checksum type 'SHA-{0}'. The only supported types are SHA1, SHA256 and SHA512"
     )]
-    InvalidChecksumType(i32),
+    InvalidChecksumType(usize),
 }
 
 #[derive(Error, Debug)]
 pub enum Sha3SumError {
     /// Error indicating that an invalid SHA3 checksum type has been provided.
     #[error("Invalid checksum type 'SHA3-{0}'. The only supported types are SHA3-224, SHA3-256, SHA3-384 and SHA3-512")]
-    InvalidChecksumType(i32),
+    InvalidChecksumType(usize),
 }
 
 #[derive(Error, Debug)]
 pub enum B2SumError {
     /// Error indicating that an invalid Blake2b checksum type has been provided.
     #[error("Invalid checksum type 'BLAKE2B-{0}'. The only supported types are BLAKE2B-256 and BLAKE2B-512")]
-    InvalidChecksumType(i32),
+    InvalidChecksumType(usize),
 }
 
 #[derive(Debug)]
